requirements.txt
fastapi
uvicorn[standard]
python-dotenv
pydantic
zen-engine